Continue troubleshooting counting results.
Something strange is going. The genes on the MT chromosomes now
have more counts. However, many of the other genes lost their
counts, including my test gene for plotting ENSG00000104904 (OAZ1).
It got almost no counts.

The Rsubread version may play a role. I ran OAZ1 in RStudio Server,
which has Rsubread 1.28.1. One BAM file alone had 23k+ assigned
reads! And this was with multimapping reads being discarded (which
is supposed to be the default).

In contrast, in the logs from my last run, which used the latest
Rsubread 2.0.0, it said in the log that multimapping reads were
allowed, and still there were almost no reads.

I'm going to re-run the data with more diagnostics to see if I
can figure out what is going on.