Only include the first read mate and count as single ended read.
The featureCounts read2pos was reducing both mates to their 5' end
and then counting both (and thus they were mapped to different bases,
and thrown out for being multi-overlapping).

https://groups.google.com/d/msg/subread/Ij5PmcHxy6I/HUd36bWMCwAJ

Thus I reduced the reads to only the first read mate and mapped as
if they are single ended. featureCounts complains that it contains
paired end reads, which I am fairly certain is because the read flags
note that they are the first in a pair.

https://davetang.org/wiki/tiki-index.php?page=SAMTools#Extracting_only_the_first_read_from_paired_end_BAM_files